Put Apollo Client in its own file; add typedefs
