+++ src/schema.tsx
@@ -1,0 +1,16 @@
+import { gql } from '@apollo/client'
+
+export const typeDefs = gql`
+  type Bookmark {
+    id: ID!
+    url: String!
+    label: String
+    description: String
+  }
+
+  type Collection {
+    id: ID!
+    title: String!
+    bookmarks: [Bookmark]
+  }
+`
